builderstream/requirements/base.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# Async Tasks
celery[redis]>=5.3,<6.0
django-celery-beat>=2.6,<3.0
redis>=5.0,<6.0

# Storage
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# Authentication
stripe>=8.0,<10.0
django-allauth>=0.61,<1.0
djangorestframework-simplejwt>=5.3,<6.0

# Environment
django-environ>=0.11,<1.0

# Media
Pillow>=10.2,<11.0
django-imagekit>=5.0,<6.0

# API Documentation
drf-spectacular>=0.27,<1.0
